Refactor lrama/parser.rb
* Separated TokenScanner class from lrama/parser.rb file
