+++ lib/lrama/parser.rb
@@ -1,4 +1,5 @@
 require "lrama/report"
+require "lrama/parser/token_scanner"
 
 module Lrama
   # Parser for parse.y, generates a grammar
@@ -6,58 +7,6 @@ module Lrama
     include Lrama::Report::Duration
 
     T = Lrama::Lexer::Token
-
-    class TokenScanner
-      def initialize(tokens)
-        @tokens = tokens
-        @index = 0
-      end
-
-      def current_token
-        @tokens[@index]
-      end
-
-      def current_type
-        current_token && current_token.type
-      end
-
-      def next
-        token = current_token
-        @index += 1
-        return token
-      end
-
-      def consume(*token_types)
-        if token_types.include?(current_type)
-          token = current_token
-          self.next
-          return token
-        end
-
-        return nil
-      end
-
-      def consume!(*token_types)
-        consume(*token_types) || (raise "#{token_types} is expected but #{current_type}. #{current_token}")
-      end
-
-      def consume_multi(*token_types)
-        a = []
-
-        while token_types.include?(current_type)
-          a << current_token
-          self.next
-        end
-
-        raise "No token is consumed. #{token_types}" if a.empty?
-
-        return a
-      end
-
-      def eots?
-        current_token.nil?
-      end
-    end
 
     def initialize(text)
       @text = text
